push bt1_QuadraticEquation

+++ src/ss04_Class_Object/thuchanh/Rectangle.java
@@ -26,7 +26,7 @@ public class Rectangle {
     }
 
     public double getArea(){
-        return (int)this.height*(int)this.width;
+        return this.height*this.width;
     }
 
     public double getPerimeter(){
